Move SASL config type to Kafka security package
We are only using SASL (and mTLS) in the context of Kafka, so should reflect this dependency and specialisation in the package structure.
