+++ scheduler/data-flow/src/main/kotlin/io/seldon/dataflow/kafka/security/Sasl.kt
@@ -1,5 +1,14 @@
 package io.seldon.dataflow.kafka.security
 
+typealias FilePath = String
+
+data class SaslConfig(
+    val mechanism: String,
+    val username: String,
+    val secret: String,
+    val passwordPath: FilePath
+)
+
 enum class KafkaSaslMechanisms(val mechanism: String) {
     PLAIN("PLAIN"),
     SCRAM_SHA_256("SCRAM-SHA-256"),
